+++ lib/jxl/modular/encoding/enc_ma.h
@@ -145,7 +145,7 @@ Status TokenizeTree(const Tree &tree, std::vector<Token> *tokens,
                     Tree *decoder_tree);
 
 void CollectPixelSamples(const Image &image, const ModularOptions &options,
-                         size_t group_id,
+                         uint32_t group_id,
                          std::vector<uint32_t> &group_pixel_count,
                          std::vector<uint32_t> &channel_pixel_count,
                          std::vector<pixel_type> &pixel_samples,
